move notify to new commands dir
